Fix incorrect filename iteration.
What was there before would keep iterating through the same iterator and try and
find files with matching names. If the `read_dir` call did not result in the
files in that directory exactly matching the order that they were written in the
file, it would fail to find them, and then prevent the model from loading.

Instead we evaluate the iterator, and then try and find each of the filenames once.