GitButler Workspace Commit
This is a merge commit of the virtual branches in your workspace.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - devin/1764483534-stargazers-hero-grid
   branch head: 134a2f752025592958b8e6f74b1bf7649f325cf8
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/branch-management/integration-branch